♻️ Rename package sqlslogopts to opts
