+++ opts/conn.go
@@ -1,4 +1,4 @@
-package sqlslogopts
+package opts
 
 import (
 	"database/sql/driver"
