+++ opts/driver_name.go
@@ -1,4 +1,4 @@
-package sqlslogopts
+package opts
 
 const (
 	DriverNameMysql = "mysql"
