+++ opts/id_gen.go
@@ -1,4 +1,4 @@
-package sqlslogopts
+package opts
 
 // IDGen is a function that generates an ID string.
 type IDGen = func() string
