+++ opts/id_gen_chacha8_test.go
@@ -1,4 +1,4 @@
-package sqlslogopts
+package opts
 
 import (
 	"fmt"
