+++ opts/level_test.go
@@ -1,4 +1,4 @@
-package sqlslogopts
+package opts
 
 import (
 	"fmt"
