+++ opts/logger.go
@@ -1,4 +1,4 @@
-package sqlslogopts
+package opts
 
 import (
 	"context"
