+++ opts/step_options.go
@@ -1,4 +1,4 @@
-package sqlslogopts
+package opts
 
 import "log/slog"
